+++ src-docs/src/routes.js
@@ -160,6 +160,8 @@ import { StatExample } from './views/stat/stat_example';
 
 import { StepsExample } from './views/steps/steps_example';
 
+// import { SuggestExample } from './views/suggest/suggest_example';
+
 import { TableExample } from './views/tables/tables_example';
 
 import { TabsExample } from './views/tabs/tabs_example';
@@ -337,6 +339,7 @@ const navigation = [
       RangeControlExample,
       SearchBarExample,
       SelectableExample,
+      // SuggestExample,
     ].map(example => createExample(example)),
   },
   {